+++ lib/doing/plugins/import/calendar_import.rb
@@ -1,38 +1,31 @@
-#!/usr/bin/env ruby
+# frozen_string_literal: true
+
 require 'json'
 
-$wwid.register_plugin({
-  name: 'calendar',
-  type: :import,
-  class: 'CalendarImport',
-  trigger: 'cal(?:endar)?'
-})
-
 class CalendarImport
-  include Util
+  include Doing::Util
   ##
   ## @brief      Imports a Timing report
   ##
   ## @param      path     (String) Path to JSON report file
   ## @param      options      (Hash) Additional Options
   ##
-  def import(path, options: {})
-    limit = 30
-    if options[:limit]
-    	limit = options[:limit].to_i
-    end
+  def import(wwid, path, options: {})
 
-    section = options[:section] || $wwid.current_section
+    limit_start = options[:start].to_i
+    limit_end = options[:end].to_i
+
+    section = options[:section] || wwid.current_section
     options[:no_overlap] ||= false
-    options[:autotag] ||= $wwid.auto_tag
+    options[:autotag] ||= wwid.auto_tag
 
-    add_section(section) unless $wwid.content.has_key?(section)
+    add_section(section) unless wwid.content.has_key?(section)
 
     tags = options[:tag] ? options[:tag].split(/[ ,]+/).map { |t| t.sub(/^@?/, '@') } : []
     prefix = options[:prefix] ? options[:prefix] : '[Calendar.app]'
 
     script = File.join(File.dirname(__FILE__), 'cal_to_json.scpt')
-    res = `/usr/bin/osascript "#{script}" #{limit}`.strip
+    res = `/usr/bin/osascript "#{script}" #{limit_start} #{limit_end}`.strip
     data = JSON.parse(res)
 
     new_items = []
@@ -54,7 +47,7 @@ class CalendarImport
           title += " @#{tag}"
         end
       end
-      title = $wwid.autotag(title) if options[:autotag]
+      title = wwid.autotag(title) if options[:autotag]
       title += " @done(#{end_time.strftime('%Y-%m-%d %H:%M')})"
       title.gsub!(/ +/, ' ')
       title.strip!
@@ -63,10 +56,17 @@ class CalendarImport
       new_items.push(new_entry)
     end
     total = new_items.count
-    new_items = $wwid.dedup(new_items, options[:no_overlap])
+    new_items = wwid.dedup(new_items, options[:no_overlap])
     dups = total - new_items.count
-    $wwid.results.push(%(Skipped #{dups} items with overlapping times)) if dups > 0
-    $wwid.content[section]['items'].concat(new_items)
-    $wwid.results.push(%(Imported #{new_items.count} items to #{section}))
+    wwid.results.push(%(Skipped #{dups} items with overlapping times)) if dups > 0
+    wwid.content[section]['items'].concat(new_items)
+    wwid.results.push(%(Imported #{new_items.count} items to #{section}))
   end
 end
+
+Doing::Plugins.register_plugin({
+  name: 'calendar',
+  type: :import,
+  class: 'CalendarImport',
+  trigger: 'cal(?:endar)?'
+})